# [perf] FillNarInfoCache during nix profile's pendingPackagesForInstallation