@savil 2023-10-04
`pendingPackagesForInstallation` runs in many different contexts, not just adding packages. Its job is to ensure the nix profile matches the devbox.json, and so it needs to check which of the packages are remaining to be added.