@savil 2023-10-04
I think we can be smarter about when to check for the HEAD request. Will investigate more...