@savil 2023-10-04
I also think merging `addPackagesToProfile` and `tidyProfile` would be beneficial